Bugfix - CR register cannot be updated properly
In the original API, the 7up CR register could *either* update whether BG/SP are showing, *OR* enable interrupts, but not both.  If one was set, the other would suffer.

This adds a eris_low_sup_setreg() function which allows to set a register's value directly, rather than piecemeal, avoiding that pitfall.